backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyJWT==2.8.0
paddleocr==3.3.1
paddlepaddle==3.2.1
spacy==3.6.1
python-dateutil==2.8.2
reportlab==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
numpy==1.26.4